make daily reset require an environment variable

+++ lib/daily_jobs.rb
@@ -1,6 +1,6 @@
 class DailyJobs
   def self.run
-    CreateTrainingDistrict.generate  if defined?(SIMS_DOMAIN) && SIMS_DOMAIN == "sims-open.vegantech.com"
+    CreateTrainingDistrict.generate  if  defined?(SIMS_DOMAIN) && SIMS_DOMAIN == "sims-open.vegantech.com"
     self.regenerate_intervention_reports
 
   end
@@ -12,22 +12,26 @@ class DailyJobs
   end
 
   def self.run_weekly
-    self.reset_demo if defined?(SIMS_DOMAIN) && SIMS_DOMAIN == "sims-open.vegantech.com"
+    self.reset_demo if ENV['RESET_DEMO'] && defined?(SIMS_DOMAIN) && SIMS_DOMAIN == "sims-open.vegantech.com"
     Notifications.setup_ending_reminders(District.find_by_state_dpi_num(3269))
   end
 
   def self.reset_demo
-    Intervention.destroy_all
-    CustomFlag.destroy_all
-    Checklist.destroy_all
-    Recommendation.destroy_all
-    PrincipalOverride.destroy_all
-    StudentComment.destroy_all
-    RailmailDelivery.destroy_all if defined?RailmailDelivery
-    ConsultationForm.destroy_all
-    ConsultationFormRequest.destroy_all
-    TeamConsultation.destroy_all
-    puts "Reset Demo"
+    if ENV['RESET_DEMO'] #being paranoid here
+      Intervention.destroy_all
+      CustomFlag.destroy_all
+      Checklist.destroy_all
+      Recommendation.destroy_all
+      PrincipalOverride.destroy_all
+      StudentComment.destroy_all
+      RailmailDelivery.destroy_all if defined?RailmailDelivery
+      ConsultationForm.destroy_all
+      ConsultationFormRequest.destroy_all
+      TeamConsultation.destroy_all
+      puts "Reset Demo"
+    else
+      puts 'ENV RESET_DEMO not set,  requiring this just to be safe'
+    end
   end
 
 end
